refactor: split codegen_emit.c into 4 files by concern
codegen_emit.c (2845 lines) was handling 5 different concerns. Split into:
- codegen_emit.c (~900 lines): output helpers, type emission, declarations
- codegen_rc.c (~140 lines): RC variable tracking
- codegen_expr.c (~1000 lines): expression emission
- codegen_stmt.c (~820 lines): statement emission
- codegen_internal.h: cross-file function declarations

Follows the existing checker.c / checker_types.c / checker_expr.c /
checker_internal.h pattern.
