+++ w0/codegen_emit.c
@@ -1,33 +1,11 @@
-#include "codegen_emit.h"
-
 #include <stdarg.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include "alloc.h"
+#include "codegen_internal.h"
 #include "types.h"
 #include "vec.h"
-
-// =============================================================================
-// RC (Reference Counting) Variable Tracking
-// =============================================================================
-//
-// The codegen tracks RC-managed variables (those created via `new`) to emit
-// automatic cleanup calls at scope exits and function returns. This acts as
-// a compile-time deterministic "GC" — each RC variable is paired with a
-// decrement function (either generic `__rc_dec` or type-specific
-// `__rc_dec_TypeName` for types with Drop or nested RC fields).
-//
-// Key functions:
-//   rc_push_var     - Register a new RC variable with its scope depth
-//   rc_cleanup_scope - Emit dec calls for vars at a given depth, remove them
-//   rc_cleanup_all   - Emit dec calls for ALL vars (before return), keep list
-//   rc_clear_all     - Free the tracking list (at function boundaries)
-//
-// rc_cleanup_scope modifies the var list (compacts it); rc_cleanup_all only
-// emits code without modifying the list. This distinction matters because
-// rc_cleanup_all is used before early returns in conditional branches where
-// the var list must remain intact for the non-returning path.
 
 // Emit indentation spaces (4 per level) at the current indent depth
 void emit_indent(CodeGen* gen) {
@@ -44,30 +22,8 @@ void emit(CodeGen* gen, const char* fmt, ...) {
     va_end(args);
 }
 
-static void emit_expr(CodeGen* gen, Node* node);
-static void emit_struct_init(CodeGen* gen, Node* node);
-static void emit_block_contents(CodeGen* gen, Node* block);
-static void emit_destruct_pattern(CodeGen* gen, DestructPattern* pattern, const char* temp_prefix,
-                                  int is_const);
-
-// Forward declarations for emit_stmt helpers
-static void emit_expr_stmt(CodeGen* gen, Node* node);
-static void emit_var_decl_stmt(CodeGen* gen, Node* node);
-static void emit_return_stmt(CodeGen* gen, Node* node);
-
-// Forward declarations for emit_expr helpers
-static void emit_string_lit(CodeGen* gen, Node* node);
-static void emit_char_lit(CodeGen* gen, Node* node);
-static void emit_string_interp(CodeGen* gen, Node* node);
-static void emit_enum_value(CodeGen* gen, Node* node);
-static void emit_call_expr(CodeGen* gen, Node* node);
-static void emit_index_expr(CodeGen* gen, Node* node);
-static void emit_slice_expr(CodeGen* gen, Node* node);
-static void emit_member_expr(CodeGen* gen, Node* node);
-static void emit_new_expr(CodeGen* gen, Node* node);
-
 // Push a deferred statement onto the defer stack for later LIFO emission
-static void defer_push(CodeGen* gen, Node* node) {
+void defer_push(CodeGen* gen, Node* node) {
     VEC_GROW(gen->defer_stack, gen->defer_count, gen->defer_capacity);
     gen->defer_stack[gen->defer_count++] = node;
 }
@@ -75,116 +31,6 @@ static void defer_push(CodeGen* gen, Node* node) {
 // Reset the defer stack (called at function boundaries)
 void defer_clear(CodeGen* gen) {
     gen->defer_count = 0;
-}
-
-// Return the appropriate __rc_dec function name for a type (type-specific or generic)
-static const char* get_dec_func_for_type(Type* t) {
-    if (t && t->kind == TYPE_STRUCT && (t->as.struc.has_drop || t->as.struc.has_rc_fields)) {
-        // Build "__rc_dec_TypeName"
-        size_t len = strlen("__rc_dec_") + strlen(t->as.struc.name) + 1;
-        char*  buf = xmalloc(len);
-        snprintf(buf, len, "__rc_dec_%s", t->as.struc.name);
-        return buf;
-    }
-    if (t && t->kind == TYPE_ENUM && t->as.enm.has_rc_fields) {
-        size_t len = strlen("__rc_dec_") + strlen(t->as.enm.name) + 1;
-        char*  buf = xmalloc(len);
-        snprintf(buf, len, "__rc_dec_%s", t->as.enm.name);
-        return buf;
-    }
-    if (t && t->kind == TYPE_VEC) {
-        const char* elem_tname = type_mangle_name(t->as.vec.elem);
-        size_t      len        = strlen("__rc_dec_Vec_") + strlen(elem_tname) + 1;
-        char*       buf        = xmalloc(len);
-        snprintf(buf, len, "__rc_dec_Vec_%s", elem_tname);
-        return buf;
-    }
-    return xstrdup("__rc_dec");
-}
-
-// Return the appropriate __rc_inc function name for a type (enum-specific or generic)
-static const char* get_inc_func_for_type(Type* t) {
-    if (t && t->kind == TYPE_ENUM && t->as.enm.has_rc_fields) {
-        size_t len = strlen("__rc_inc_") + strlen(t->as.enm.name) + 1;
-        char*  buf = xmalloc(len);
-        snprintf(buf, len, "__rc_inc_%s", t->as.enm.name);
-        return buf;
-    }
-    return xstrdup("__rc_inc");
-}
-
-// Register an RC-managed variable for scope-based cleanup tracking
-static void rc_push_var(CodeGen* gen, const char* name, const char* dec_func, Type* type) {
-    VEC_GROW(gen->rc_vars, gen->rc_var_count, gen->rc_var_capacity);
-    gen->rc_vars[gen->rc_var_count].name        = xstrdup(name);
-    gen->rc_vars[gen->rc_var_count].dec_func    = xstrdup(dec_func);
-    gen->rc_vars[gen->rc_var_count].type        = type;
-    gen->rc_vars[gen->rc_var_count].scope_depth = gen->rc_scope_depth;
-    gen->rc_var_count++;
-}
-
-// Emit dec for vars at the given depth, remove them from list
-static void rc_cleanup_scope(CodeGen* gen, int depth) {
-    int dst = 0;
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        if (gen->rc_vars[i].scope_depth == depth) {
-            emit_indent(gen);
-            emit(gen, "%s(%s);\n", gen->rc_vars[i].dec_func, gen->rc_vars[i].name);
-            free(gen->rc_vars[i].name);
-            free(gen->rc_vars[i].dec_func);
-        } else {
-            gen->rc_vars[dst++] = gen->rc_vars[i];
-        }
-    }
-    gen->rc_var_count = dst;
-}
-
-// Emit dec for ALL remaining vars (skip one by name). Does NOT modify list.
-static void rc_cleanup_all(CodeGen* gen, const char* skip_name) {
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        if (skip_name && strcmp(gen->rc_vars[i].name, skip_name) == 0) {
-            continue;
-        }
-        emit_indent(gen);
-        emit(gen, "%s(%s);\n", gen->rc_vars[i].dec_func, gen->rc_vars[i].name);
-    }
-}
-
-// Free and reset the RC var list (at function boundary)
-void rc_clear_all(CodeGen* gen) {
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        free(gen->rc_vars[i].name);
-        free(gen->rc_vars[i].dec_func);
-    }
-    gen->rc_var_count   = 0;
-    gen->rc_scope_depth = 0;
-}
-
-// Look up the stored dec_func for a tracked RC variable
-static const char* rc_get_dec_func(CodeGen* gen, const char* name) {
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        if (strcmp(gen->rc_vars[i].name, name) == 0)
-            return gen->rc_vars[i].dec_func;
-    }
-    return "__rc_dec";
-}
-
-// Look up the stored Type* for a tracked RC variable
-static Type* rc_get_var_type(CodeGen* gen, const char* name) {
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        if (strcmp(gen->rc_vars[i].name, name) == 0)
-            return gen->rc_vars[i].type;
-    }
-    return NULL;
-}
-
-// Check if a variable name is in the RC tracking list
-static int rc_is_tracked(CodeGen* gen, const char* name) {
-    for (int i = 0; i < gen->rc_var_count; i++) {
-        if (strcmp(gen->rc_vars[i].name, name) == 0)
-            return 1;
-    }
-    return 0;
 }
 
 // Check if a name is a registered enum type
@@ -216,7 +62,7 @@ int enum_has_rc_fields(CodeGen* gen, const char* name) {
 
 // Resolve a type node through aliases. If the node is a NODE_IDENT
 // that names a type alias, return the alias target node instead.
-static Node* resolve_alias(CodeGen* gen, Node* type_node) {
+Node* resolve_alias(CodeGen* gen, Node* type_node) {
     if (!type_node || type_node->type != NODE_IDENT)
         return type_node;
     for (int i = 0; i < gen->alias_count; i++) {
@@ -228,7 +74,7 @@ static Node* resolve_alias(CodeGen* gen, Node* type_node) {
 }
 
 // Check if a type node represents a struct (user-defined) type
-static int is_struct_type(CodeGen* gen, Node* type_node) {
+int is_struct_type(CodeGen* gen, Node* type_node) {
     if (!type_node)
         return 0;
     type_node = resolve_alias(gen, type_node);
@@ -752,7 +598,7 @@ void emit_type_with_name(CodeGen* gen, Node* type_node, const char* name) {
 }
 
 // Map a binary operator token to its C operator string
-static const char* binary_op_str(TokenType op) {
+const char* binary_op_str(TokenType op) {
     switch (op) {
     case TOK_PLUS:
         return "+";
@@ -796,7 +642,7 @@ static const char* binary_op_str(TokenType op) {
 }
 
 // Map a unary operator token to its C operator string
-static const char* unary_op_str(TokenType op) {
+const char* unary_op_str(TokenType op) {
     switch (op) {
     case TOK_MINUS:
         return "-";
@@ -814,7 +660,7 @@ static const char* unary_op_str(TokenType op) {
 }
 
 // Map an assignment operator token to its C operator string
-static const char* assign_op_str(TokenType op) {
+const char* assign_op_str(TokenType op) {
     switch (op) {
     case TOK_EQ:
         return "=";
@@ -840,1795 +686,6 @@ static const char* assign_op_str(TokenType op) {
         return ">>=";
     default:
         return "=";
-    }
-}
-
-// In a generic method body, look up the field type node from the struct template.
-// For `self.fieldname`, returns the AST type node of the field, or NULL.
-static Node* lookup_generic_template_field_type(CodeGen* gen, const char* field_name) {
-    if (!gen->current_generic_template)
-        return NULL;
-    Node* tmpl = gen->current_generic_template;
-    for (int i = 0; i < tmpl->as.struct_decl.fields.count; i++) {
-        Node* field = tmpl->as.struct_decl.fields.nodes[i];
-        if (strcmp(field->as.field.name, field_name) == 0) {
-            return field->as.field.type;
-        }
-    }
-    return NULL;
-}
-
-// Resolve the struct name for a method call in a generic method body.
-// For `self.field.method(...)`, determines if `field` is a Vec/struct and returns
-// the mangled name (e.g., "__Vec_HashEntry_i32" or "HashEntry_i32").
-// Returns a malloc'd string, or NULL if unresolvable. Caller must free.
-static char* resolve_generic_method_target(CodeGen* gen, Node* member) {
-    if (!gen->subst_ctx || !gen->current_generic_template)
-        return NULL;
-    // Only handle `self.field.method()` pattern
-    Node* obj = member->as.member.object;
-    if (!obj || obj->type != NODE_MEMBER)
-        return NULL;
-    // obj should be `self.field`
-    if (!obj->as.member.object || obj->as.member.object->type != NODE_IDENT)
-        return NULL;
-    if (strcmp(obj->as.member.object->as.ident.name, "self") != 0)
-        return NULL;
-    // Look up the field type in the template
-    Node* field_type = lookup_generic_template_field_type(gen, obj->as.member.name);
-    if (!field_type || field_type->type != NODE_GENERIC_TYPE)
-        return NULL;
-    const char* base = field_type->as.generic_type.base_name;
-    if (strcmp(base, "Vec") == 0) {
-        // Build "__Vec_" + mangled element type
-        Node* elem = field_type->as.generic_type.type_args.nodes[0];
-        char  buf[256];
-        if (elem->type == NODE_IDENT) {
-            const char* elem_name = elem->as.ident.name;
-            // Check substitution
-            if (gen->subst_ctx) {
-                for (int s = 0; s < gen->subst_ctx->count; s++) {
-                    if (strcmp(gen->subst_ctx->type_params[s], elem_name) == 0) {
-                        snprintf(buf, sizeof(buf), "__Vec_%s",
-                                 type_mangle_name(gen->subst_ctx->type_args[s]));
-                        return xstrdup(buf);
-                    }
-                }
-            }
-            snprintf(buf, sizeof(buf), "__Vec_%s", elem_name);
-            return xstrdup(buf);
-        } else if (elem->type == NODE_GENERIC_TYPE) {
-            char* mangled = build_mangled_name_from_generic_node(gen, elem);
-            snprintf(buf, sizeof(buf), "__Vec_%s", mangled);
-            free(mangled);
-            return xstrdup(buf);
-        }
-        return NULL;
-    }
-    // Generic struct type — build mangled name
-    char* mangled = build_mangled_name_from_generic_node(gen, field_type);
-    return mangled;
-}
-
-// Build the __rc_dec function name for a field type AST node in a generic method body.
-// Returns a malloc'd string, or NULL if the field is not RC-managed. Caller must free.
-static char* resolve_generic_field_dec_func(CodeGen* gen, Node* field_type_node) {
-    if (!field_type_node)
-        return NULL;
-    if (field_type_node->type == NODE_GENERIC_TYPE) {
-        const char* base = field_type_node->as.generic_type.base_name;
-        if (strcmp(base, "Vec") == 0) {
-            // Vec<T> → __rc_dec_Vec_T
-            Node* elem = field_type_node->as.generic_type.type_args.nodes[0];
-            char  buf[256];
-            if (elem->type == NODE_IDENT) {
-                const char* elem_name = elem->as.ident.name;
-                if (gen->subst_ctx) {
-                    for (int s = 0; s < gen->subst_ctx->count; s++) {
-                        if (strcmp(gen->subst_ctx->type_params[s], elem_name) == 0) {
-                            snprintf(buf, sizeof(buf), "__rc_dec_Vec_%s",
-                                     type_name(gen->subst_ctx->type_args[s]));
-                            return xstrdup(buf);
-                        }
-                    }
-                }
-                snprintf(buf, sizeof(buf), "__rc_dec_Vec_%s", elem_name);
-                return xstrdup(buf);
-            } else if (elem->type == NODE_GENERIC_TYPE) {
-                char* mangled = build_mangled_name_from_generic_node(gen, elem);
-                snprintf(buf, sizeof(buf), "__rc_dec_Vec_%s", mangled);
-                free(mangled);
-                return xstrdup(buf);
-            }
-            return NULL;
-        }
-        if (strcmp(base, "Span") == 0)
-            return NULL; // Spans are value types, not RC
-        // Generic struct → __rc_dec_MangledName
-        char* mangled = build_mangled_name_from_generic_node(gen, field_type_node);
-        char  buf[256];
-        snprintf(buf, sizeof(buf), "__rc_dec_%s", mangled);
-        free(mangled);
-        return xstrdup(buf);
-    }
-    if (field_type_node->type == NODE_IDENT) {
-        const char* name = field_type_node->as.ident.name;
-        // Check if it's a type param that resolves to a struct
-        if (gen->subst_ctx) {
-            for (int s = 0; s < gen->subst_ctx->count; s++) {
-                if (strcmp(gen->subst_ctx->type_params[s], name) == 0) {
-                    Type* resolved = gen->subst_ctx->type_args[s];
-                    if (resolved->kind == TYPE_STRUCT)
-                        return (char*)get_dec_func_for_type(resolved);
-                    return NULL;
-                }
-            }
-        }
-        // Non-generic struct field
-        if (!type_is_builtin_name(name) && !is_enum_type_name(gen, name)) {
-            char buf[256];
-            snprintf(buf, sizeof(buf), "__rc_dec_%s", name);
-            return xstrdup(buf);
-        }
-    }
-    return NULL;
-}
-
-// Emit a string literal with C escape sequences
-static void emit_string_lit(CodeGen* gen, Node* node) {
-    emit(gen, "\"");
-    // Escape special characters
-    for (int i = 0; i < node->as.string_lit.length; i++) {
-        char c = node->as.string_lit.value[i];
-        switch (c) {
-        case '\n':
-            emit(gen, "\\n");
-            break;
-        case '\t':
-            emit(gen, "\\t");
-            break;
-        case '\r':
-            emit(gen, "\\r");
-            break;
-        case '\\':
-            emit(gen, "\\\\");
-            break;
-        case '"':
-            emit(gen, "\\\"");
-            break;
-        default:
-            emit(gen, "%c", c);
-            break;
-        }
-    }
-    emit(gen, "\"");
-}
-
-// Emit a character literal with C escape sequences
-static void emit_char_lit(CodeGen* gen, Node* node) {
-    if (node->as.char_lit.value == '\n') {
-        emit(gen, "'\\n'");
-    } else if (node->as.char_lit.value == '\t') {
-        emit(gen, "'\\t'");
-    } else if (node->as.char_lit.value == '\r') {
-        emit(gen, "'\\r'");
-    } else if (node->as.char_lit.value == '\\') {
-        emit(gen, "'\\\\'");
-    } else if (node->as.char_lit.value == '\'') {
-        emit(gen, "'\\''");
-    } else {
-        emit(gen, "'%c'", node->as.char_lit.value);
-    }
-}
-
-// Emit an enum variant: simple tag, bare data enum, or data enum with constructor args
-static void emit_enum_value(CodeGen* gen, Node* node) {
-    if (!node->as.enum_value.is_data_enum) {
-        // Simple enum: emit qualified value name (EnumName_ValueName)
-        emit(gen, "%.*s_%.*s", node->as.enum_value.enum_name_length, node->as.enum_value.enum_name,
-             node->as.enum_value.value_name_length, node->as.enum_value.value_name);
-    } else if (node->as.enum_value.args.count == 0) {
-        // Data enum, bare tag: (EnumName){.tag = EnumName_ValueName}
-        emit(gen, "(%.*s){.tag = %.*s_%.*s}", node->as.enum_value.enum_name_length,
-             node->as.enum_value.enum_name, node->as.enum_value.enum_name_length,
-             node->as.enum_value.enum_name, node->as.enum_value.value_name_length,
-             node->as.enum_value.value_name);
-    } else {
-        // Data enum with args: (EnumName){.tag = EnumName_ValueName, .ValueName = {.f0 = ..}}
-        int needs_rc_inc = 0;
-        for (int i = 0; i < node->as.enum_value.args.count; i++) {
-            Node* arg = node->as.enum_value.args.nodes[i];
-            if (arg->type == NODE_IDENT && rc_is_tracked(gen, arg->as.ident.name)) {
-                needs_rc_inc = 1;
-                break;
-            }
-        }
-
-        if (needs_rc_inc) {
-            emit(gen, "({ ");
-            for (int i = 0; i < node->as.enum_value.args.count; i++) {
-                Node* arg = node->as.enum_value.args.nodes[i];
-                if (arg->type == NODE_IDENT && rc_is_tracked(gen, arg->as.ident.name)) {
-                    Type*       arg_type = rc_get_var_type(gen, arg->as.ident.name);
-                    const char* inc_fn   = get_inc_func_for_type(arg_type);
-                    emit(gen, "%s(%s); ", inc_fn, arg->as.ident.name);
-                    free((char*)inc_fn);
-                }
-            }
-        }
-
-        emit(gen, "(%.*s){.tag = %.*s_%.*s, .%.*s = {", node->as.enum_value.enum_name_length,
-             node->as.enum_value.enum_name, node->as.enum_value.enum_name_length,
-             node->as.enum_value.enum_name, node->as.enum_value.value_name_length,
-             node->as.enum_value.value_name, node->as.enum_value.value_name_length,
-             node->as.enum_value.value_name);
-        for (int i = 0; i < node->as.enum_value.args.count; i++) {
-            if (i > 0)
-                emit(gen, ", ");
-            emit(gen, ".f%d = ", i);
-            emit_expr(gen, node->as.enum_value.args.nodes[i]);
-        }
-        emit(gen, "}}");
-
-        if (needs_rc_inc) {
-            emit(gen, "; })");
-        }
-    }
-}
-
-// Emit a function call: module-qualified, method, generic method, or regular call
-static void emit_call_expr(CodeGen* gen, Node* node) {
-    Node* func = node->as.call.func;
-    // Check if this is a module-qualified call (e.g., std.print())
-    if (func->type == NODE_MEMBER && func->as.member.module_name != NULL) {
-        // Built-in: std.format -> __std_format
-        if (strcmp(func->as.member.module_name, "std") == 0 &&
-            strncmp(func->as.member.name, "format", func->as.member.length) == 0 &&
-            func->as.member.length == 6) {
-            emit(gen, "__std_format(");
-            for (int i = 0; i < node->as.call.args.count; i++) {
-                if (i > 0)
-                    emit(gen, ", ");
-                emit_expr(gen, node->as.call.args.nodes[i]);
-            }
-            emit(gen, ")");
-            return;
-        }
-        // Module-qualified call: emit module_func(args...)
-        emit(gen, "%s_%.*s(", func->as.member.module_name, func->as.member.length,
-             func->as.member.name);
-        for (int i = 0; i < node->as.call.args.count; i++) {
-            if (i > 0)
-                emit(gen, ", ");
-            emit_expr(gen, node->as.call.args.nodes[i]);
-        }
-        emit(gen, ")");
-    } else if (func->type == NODE_MEMBER && func->as.member.struct_name != NULL) {
-        // Method call: emit StructName_method(obj, args...)
-        // With struct references, objects are already pointers
-        emit(gen, "%s_%.*s(", func->as.member.struct_name, func->as.member.length,
-             func->as.member.name);
-        // Emit the receiver as first argument (already a pointer)
-        emit_expr(gen, func->as.member.object);
-        // Emit remaining arguments
-        for (int i = 0; i < node->as.call.args.count; i++) {
-            emit(gen, ", ");
-            emit_expr(gen, node->as.call.args.nodes[i]);
-        }
-        emit(gen, ")");
-    } else if (func->type == NODE_MEMBER && func->as.member.struct_name == NULL && gen->subst_ctx) {
-        // In a generic method body — checker didn't annotate struct_name or module_name.
-        // First check if this is a module-qualified call (e.g., std.print)
-        int is_module_call = 0;
-        if (func->as.member.object->type == NODE_IDENT) {
-            const char* obj_name = func->as.member.object->as.ident.name;
-            for (int i = 0; i < gen->accessible_modules_count; i++) {
-                if (strcmp(gen->accessible_modules[i], obj_name) == 0) {
-                    is_module_call = 1;
-                    break;
-                }
-            }
-        }
-        if (is_module_call) {
-            // Module-qualified call: emit module_func(args...)
-            emit(gen, "%s_%.*s(", func->as.member.object->as.ident.name, func->as.member.length,
-                 func->as.member.name);
-            for (int i = 0; i < node->as.call.args.count; i++) {
-                if (i > 0)
-                    emit(gen, ", ");
-                emit_expr(gen, node->as.call.args.nodes[i]);
-            }
-            emit(gen, ")");
-        } else {
-            // Try to resolve the target type from the struct template field types.
-            char* resolved_name = resolve_generic_method_target(gen, func);
-            if (resolved_name) {
-                emit(gen, "%s_%.*s(", resolved_name, func->as.member.length, func->as.member.name);
-                emit_expr(gen, func->as.member.object);
-                for (int i = 0; i < node->as.call.args.count; i++) {
-                    emit(gen, ", ");
-                    emit_expr(gen, node->as.call.args.nodes[i]);
-                }
-                emit(gen, ")");
-                free(resolved_name);
-            } else {
-                // Fallback: regular function call
-                emit_expr(gen, func);
-                emit(gen, "(");
-                for (int i = 0; i < node->as.call.args.count; i++) {
-                    if (i > 0)
-                        emit(gen, ", ");
-                    emit_expr(gen, node->as.call.args.nodes[i]);
-                }
-                emit(gen, ")");
-            }
-        }
-    } else {
-        // Regular function call — check for extern alias
-        int alias_found = 0;
-        if (func->type == NODE_IDENT) {
-            for (int i = 0; i < gen->extern_alias_count; i++) {
-                if (strncmp(gen->extern_aliases[i].whist_name, func->as.ident.name,
-                            func->as.ident.length) == 0 &&
-                    gen->extern_aliases[i].whist_name[func->as.ident.length] == '\0') {
-                    emit(gen, "%s(", gen->extern_aliases[i].c_name);
-                    alias_found = 1;
-                    break;
-                }
-            }
-        }
-        // Check use aliases (use std.print -> std_print)
-        if (!alias_found && func->type == NODE_IDENT) {
-            for (int i = 0; i < gen->use_alias_count; i++) {
-                if (strncmp(gen->use_aliases[i].whist_name, func->as.ident.name,
-                            func->as.ident.length) == 0 &&
-                    gen->use_aliases[i].whist_name[func->as.ident.length] == '\0') {
-                    emit(gen, "%s(", gen->use_aliases[i].c_name);
-                    alias_found = 1;
-                    break;
-                }
-            }
-        }
-        if (!alias_found) {
-            // If inside a module, prefix non-extern function calls with module name
-            if (gen->current_module && func->type == NODE_IDENT) {
-                int is_extern = 0;
-                for (int i = 0; i < gen->extern_func_count; i++) {
-                    if (strncmp(gen->extern_funcs[i], func->as.ident.name, func->as.ident.length) ==
-                            0 &&
-                        gen->extern_funcs[i][func->as.ident.length] == '\0') {
-                        is_extern = 1;
-                        break;
-                    }
-                }
-                if (!is_extern) {
-                    emit(gen, "%s_", gen->current_module);
-                }
-            }
-            emit_expr(gen, func);
-            emit(gen, "(");
-        }
-        for (int i = 0; i < node->as.call.args.count; i++) {
-            if (i > 0)
-                emit(gen, ", ");
-            emit_expr(gen, node->as.call.args.nodes[i]);
-        }
-        emit(gen, ")");
-    }
-}
-
-// Emit an index expression: bounds-checked vec/span, tuple field, or array access
-static void emit_index_expr(CodeGen* gen, Node* node) {
-    if (node->as.index.is_vec_index) {
-        // Bounds-checked vec access
-        emit(gen, "(__w0_vec_check(");
-        emit_expr(gen, node->as.index.object);
-        emit(gen, "->count, ");
-        emit_expr(gen, node->as.index.index);
-        emit(gen, ", %d, %d), ", node->line, node->column);
-        emit_expr(gen, node->as.index.object);
-        emit(gen, "->data[");
-        emit_expr(gen, node->as.index.index);
-        emit(gen, "])");
-    } else if (node->as.index.is_span_index) {
-        // Bounds-checked span access
-        emit(gen, "(__w0_span_check(");
-        emit_expr(gen, node->as.index.object);
-        emit(gen, ".count, ");
-        emit_expr(gen, node->as.index.index);
-        emit(gen, ", %d, %d), ", node->line, node->column);
-        emit_expr(gen, node->as.index.object);
-        emit(gen, ".data[");
-        emit_expr(gen, node->as.index.index);
-        emit(gen, "])");
-    } else if (node->as.index.is_tuple_index) {
-        // Tuple indexing: obj._N
-        emit_expr(gen, node->as.index.object);
-        emit(gen, "._%ld", node->as.index.index->as.int_lit.value);
-    } else {
-        // Array/string indexing: obj[index]
-        emit_expr(gen, node->as.index.object);
-        emit(gen, "[");
-        emit_expr(gen, node->as.index.index);
-        emit(gen, "]");
-    }
-}
-
-// Emit a slice expression as a Span compound literal with data pointer and count
-static void emit_slice_expr(CodeGen* gen, Node* node) {
-    // String slicing: produces a new string via __String_substr
-    if (node->as.slice.is_string) {
-        emit(gen, "__String_substr(");
-        emit_expr(gen, node->as.slice.object);
-        emit(gen, ", ");
-        if (node->as.slice.start) {
-            emit_expr(gen, node->as.slice.start);
-        } else {
-            emit(gen, "0");
-        }
-        emit(gen, ", ");
-        if (node->as.slice.end) {
-            emit_expr(gen, node->as.slice.end);
-        } else {
-            emit(gen, "(int64_t)strlen(");
-            emit_expr(gen, node->as.slice.object);
-            emit(gen, ")");
-        }
-        emit(gen, ")");
-        return;
-    }
-
-    // Slice produces a span: (__Span_T){ .data = ..., .count = ... }
-    Type* span_type = node->as.slice.resolved_type;
-    Type* elem_type = span_type->as.span.elem;
-
-    // Emit compound literal
-    emit(gen, "((__Span_%s){ .data = ", type_mangle_name(elem_type));
-
-    if (node->as.slice.is_vec) {
-        // Vec slicing: .data = vec->data + start
-        emit_expr(gen, node->as.slice.object);
-        emit(gen, "->data + ");
-        if (node->as.slice.start) {
-            emit_expr(gen, node->as.slice.start);
-        } else {
-            emit(gen, "0");
-        }
-    } else if (node->as.slice.is_array) {
-        // Array slicing: .data = &arr[start]
-        emit(gen, "&(");
-        emit_expr(gen, node->as.slice.object);
-        emit(gen, ")[");
-        if (node->as.slice.start) {
-            emit_expr(gen, node->as.slice.start);
-        } else {
-            emit(gen, "0");
-        }
-        emit(gen, "]");
-    } else {
-        // Span slicing: .data = span.data + start
-        emit_expr(gen, node->as.slice.object);
-        emit(gen, ".data + ");
-        if (node->as.slice.start) {
-            emit_expr(gen, node->as.slice.start);
-        } else {
-            emit(gen, "0");
-        }
-    }
-
-    emit(gen, ", .count = ");
-
-    // Calculate count: end - start
-    // For omitted end, use array length or span/vec.count
-    if (node->as.slice.end) {
-        emit(gen, "(");
-        emit_expr(gen, node->as.slice.end);
-        emit(gen, ")");
-    } else {
-        // Use full length
-        if (node->as.slice.is_vec) {
-            emit_expr(gen, node->as.slice.object);
-            emit(gen, "->count");
-        } else if (node->as.slice.is_array) {
-            emit(gen, "(sizeof(");
-            emit_expr(gen, node->as.slice.object);
-            emit(gen, ")/sizeof((");
-            emit_expr(gen, node->as.slice.object);
-            emit(gen, ")[0]))");
-        } else {
-            emit_expr(gen, node->as.slice.object);
-            emit(gen, ".count");
-        }
-    }
-
-    // Subtract start if present
-    if (node->as.slice.start) {
-        emit(gen, " - (");
-        emit_expr(gen, node->as.slice.start);
-        emit(gen, ")");
-    }
-
-    emit(gen, " })");
-}
-
-// Emit a member access expression using -> for struct pointers or . for value types
-static void emit_member_expr(CodeGen* gen, Node* node) {
-    // Check if this is module-qualified access (already handled struct_name case)
-    if (node->as.member.struct_name == NULL && node->as.member.module_name == NULL) {
-        // Check if object is 'self' - always a pointer in methods
-        // This handles generic methods where is_ref isn't set because body isn't type-checked
-        int is_self = (node->as.member.object->type == NODE_IDENT &&
-                       strcmp(node->as.member.object->as.ident.name, "self") == 0);
-
-        if (node->as.member.is_ref || is_self) {
-            // Struct reference or self - use ->
-            emit_expr(gen, node->as.member.object);
-            emit(gen, "->%.*s", node->as.member.length, node->as.member.name);
-        } else {
-            // Value type member access (tuples, spans) - use .
-            emit_expr(gen, node->as.member.object);
-            emit(gen, ".%.*s", node->as.member.length, node->as.member.name);
-        }
-    } else {
-        // Struct method or module access
-        emit_expr(gen, node->as.member.object);
-        emit(gen, "->%.*s", node->as.member.length, node->as.member.name);
-    }
-}
-
-// Emit a `new` expression as a GCC statement expression: __rc_alloc + field init
-static void emit_new_expr(CodeGen* gen, Node* node) {
-    Type* rtype = node->as.new_expr.resolved_type;
-    // In generic method bodies, the checker doesn't visit the body, so resolved_type
-    // may be NULL. Resolve from the type_node using the current substitution context.
-    if (!rtype) {
-        Node* tn = node->as.new_expr.type_node;
-        if (tn->type == NODE_GENERIC_TYPE) {
-            if (strcmp(tn->as.generic_type.base_name, "Vec") == 0) {
-                // Look up the Vec instance by mangled name
-                char* mangled = build_mangled_name_from_generic_node(gen, tn);
-                for (int vi = 0; vi < gen->vec_instance_count; vi++) {
-                    if (strcmp(gen->vec_instances[vi].mangled_name, mangled) == 0) {
-                        rtype = gen->vec_instances[vi].type;
-                        break;
-                    }
-                }
-                free(mangled);
-            } else {
-                // Look up the generic struct instance by mangled name
-                char* mangled = build_mangled_name_from_generic_node(gen, tn);
-                for (int gi = 0; gi < gen->generic_instance_count; gi++) {
-                    if (strcmp(gen->generic_instances[gi].mangled_name, mangled) == 0) {
-                        rtype = gen->generic_instances[gi].type;
-                        break;
-                    }
-                }
-                free(mangled);
-            }
-        } else if (tn->type == NODE_IDENT) {
-            // Simple type name — check substitution context first
-            const char* name = tn->as.ident.name;
-            if (gen->subst_ctx) {
-                for (int si = 0; si < gen->subst_ctx->count; si++) {
-                    if (strcmp(gen->subst_ctx->type_params[si], name) == 0) {
-                        rtype = gen->subst_ctx->type_args[si];
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    if (rtype->kind == TYPE_VEC) {
-        // new Vec<T>{elems} as inline expression using GCC statement expression
-        const char* elem_tname = type_mangle_name(rtype->as.vec.elem);
-        int         tmp        = gen->temp_count++;
-        emit(gen,
-             "({ __Vec_%s* __rc_tmp%d = (__Vec_%s*)__rc_alloc(sizeof(__Vec_%s)); "
-             "__rc_tmp%d->data = NULL; __rc_tmp%d->count = 0; __rc_tmp%d->capacity = 0;",
-             elem_tname, tmp, elem_tname, elem_tname, tmp, tmp, tmp);
-        // Push initial elements
-        Node* init = node->as.new_expr.init;
-        for (int i = 0; i < init->as.struct_init.fields.count; i++) {
-            Node* field = init->as.struct_init.fields.nodes[i];
-            if (field && field->type == NODE_FIELD_INIT) {
-                emit(gen, " __Vec_%s_push(__rc_tmp%d, ", elem_tname, tmp);
-                emit_expr(gen, field->as.field_init.value);
-                emit(gen, ");");
-            }
-        }
-        emit(gen, " __rc_tmp%d; })", tmp);
-    } else {
-        // new Type { fields } as inline expression using GCC statement expression
-        const char* tname = rtype->as.struc.name;
-        int         tmp   = gen->temp_count++;
-        emit(gen, "({ %s* __rc_tmp%d = (%s*)__rc_alloc(sizeof(%s)); *__rc_tmp%d = (%s)", tname, tmp,
-             tname, tname, tmp, tname);
-        emit_struct_init(gen, node->as.new_expr.init);
-        emit(gen, ";");
-        // Increment refcount for any RC-tracked idents stored in struct fields
-        Node* rc_init = node->as.new_expr.init;
-        for (int i = 0; i < rc_init->as.struct_init.fields.count; i++) {
-            Node* field = rc_init->as.struct_init.fields.nodes[i];
-            if (field && field->type == NODE_FIELD_INIT &&
-                field->as.field_init.value->type == NODE_IDENT &&
-                rc_is_tracked(gen, field->as.field_init.value->as.ident.name)) {
-                const char* vname  = field->as.field_init.value->as.ident.name;
-                Type*       vtype  = rc_get_var_type(gen, vname);
-                const char* inc_fn = get_inc_func_for_type(vtype);
-                emit(gen, " %s(%s);", inc_fn, vname);
-                free((char*)inc_fn);
-            }
-        }
-        emit(gen, " __rc_tmp%d; })", tmp);
-    }
-}
-
-// Emit string interpolation: $"text {expr} text" -> __std_format("fmt", args...)
-static void emit_string_interp(CodeGen* gen, Node* node) {
-    int count = node->as.string_interp.part_count;
-
-    // Optimization: if no expression parts, emit as a plain C string literal
-    int has_expr = 0;
-    for (int i = 0; i < count; i++) {
-        if (node->as.string_interp.parts.nodes[i]->type != NODE_STRING_LIT) {
-            has_expr = 1;
-            break;
-        }
-    }
-
-    if (!has_expr) {
-        // All parts are text — concatenate into a single string literal
-        emit(gen, "\"");
-        for (int i = 0; i < count; i++) {
-            Node* part = node->as.string_interp.parts.nodes[i];
-            // Emit with C escaping
-            const char* s = part->as.string_lit.value;
-            int         n = part->as.string_lit.length;
-            for (int j = 0; j < n; j++) {
-                switch (s[j]) {
-                case '\n':
-                    emit(gen, "\\n");
-                    break;
-                case '\t':
-                    emit(gen, "\\t");
-                    break;
-                case '\r':
-                    emit(gen, "\\r");
-                    break;
-                case '\\':
-                    emit(gen, "\\\\");
-                    break;
-                case '"':
-                    emit(gen, "\\\"");
-                    break;
-                case '\0':
-                    emit(gen, "\\0");
-                    break;
-                default:
-                    emit(gen, "%c", s[j]);
-                    break;
-                }
-            }
-        }
-        emit(gen, "\"");
-        return;
-    }
-
-    // Build __std_format("fmt", args...)
-    emit(gen, "__std_format(\"");
-
-    // First pass: format string
-    for (int i = 0; i < count; i++) {
-        Node* part = node->as.string_interp.parts.nodes[i];
-        Type* t    = node->as.string_interp.part_types[i];
-        if (part->type == NODE_STRING_LIT) {
-            // Text segment: emit with C escaping, and escape % as %%
-            const char* s = part->as.string_lit.value;
-            int         n = part->as.string_lit.length;
-            for (int j = 0; j < n; j++) {
-                switch (s[j]) {
-                case '\n':
-                    emit(gen, "\\n");
-                    break;
-                case '\t':
-                    emit(gen, "\\t");
-                    break;
-                case '\r':
-                    emit(gen, "\\r");
-                    break;
-                case '\\':
-                    emit(gen, "\\\\");
-                    break;
-                case '"':
-                    emit(gen, "\\\"");
-                    break;
-                case '\0':
-                    emit(gen, "\\0");
-                    break;
-                case '%':
-                    emit(gen, "%%%%");
-                    break;
-                default:
-                    emit(gen, "%c", s[j]);
-                    break;
-                }
-            }
-        } else {
-            // Expression: emit format specifier based on type
-            if (!t)
-                continue;
-            switch (t->kind) {
-            case TYPE_STRING:
-                emit(gen, "%%s");
-                break;
-            case TYPE_BOOL:
-                emit(gen, "%%s");
-                break;
-            case TYPE_CHAR:
-                emit(gen, "%%c");
-                break;
-            case TYPE_F32:
-            case TYPE_F64:
-                emit(gen, "%%g");
-                break;
-            case TYPE_INT64:
-                emit(gen, "%%lld");
-                break;
-            case TYPE_UINT64:
-                emit(gen, "%%llu");
-                break;
-            default:
-                // Other integer types
-                emit(gen, "%%d");
-                break;
-            }
-        }
-    }
-    emit(gen, "\"");
-
-    // Second pass: arguments
-    for (int i = 0; i < count; i++) {
-        Node* part = node->as.string_interp.parts.nodes[i];
-        Type* t    = node->as.string_interp.part_types[i];
-        if (part->type == NODE_STRING_LIT)
-            continue;
-        if (!t)
-            continue;
-        emit(gen, ", ");
-        if (t->kind == TYPE_BOOL) {
-            emit(gen, "(");
-            emit_expr(gen, part);
-            emit(gen, " ? \"true\" : \"false\")");
-        } else {
-            emit_expr(gen, part);
-        }
-    }
-
-    emit(gen, ")");
-}
-
-// Dispatch expression code generation based on node type
-static void emit_expr(CodeGen* gen, Node* node) {
-    if (!node)
-        return;
-
-    switch (node->type) {
-    case NODE_INT_LIT:
-        emit(gen, "%ldLL", node->as.int_lit.value);
-        break;
-
-    case NODE_FLOAT_LIT:
-        emit(gen, "%g", node->as.float_lit.value);
-        break;
-
-    case NODE_STRING_LIT:
-        emit_string_lit(gen, node);
-        break;
-
-    case NODE_CHAR_LIT:
-        emit_char_lit(gen, node);
-        break;
-
-    case NODE_BOOL_LIT:
-        emit(gen, "%s", node->as.bool_lit.value ? "true" : "false");
-        break;
-
-    case NODE_NULL_LIT:
-        emit(gen, "NULL");
-        break;
-
-    case NODE_IDENT:
-        emit(gen, "%.*s", node->as.ident.length, node->as.ident.name);
-        break;
-
-    case NODE_ENUM_VALUE:
-        emit_enum_value(gen, node);
-        break;
-
-    case NODE_BINARY:
-        if (node->as.binary.is_string_op) {
-            TokenType op = node->as.binary.op;
-            if (op == TOK_EQ_EQ) {
-                emit(gen, "(strcmp(");
-                emit_expr(gen, node->as.binary.left);
-                emit(gen, ", ");
-                emit_expr(gen, node->as.binary.right);
-                emit(gen, ") == 0)");
-            } else if (op == TOK_BANG_EQ) {
-                emit(gen, "(strcmp(");
-                emit_expr(gen, node->as.binary.left);
-                emit(gen, ", ");
-                emit_expr(gen, node->as.binary.right);
-                emit(gen, ") != 0)");
-            } else if (op == TOK_PLUS) {
-                emit(gen, "__String_concat(");
-                emit_expr(gen, node->as.binary.left);
-                emit(gen, ", ");
-                emit_expr(gen, node->as.binary.right);
-                emit(gen, ")");
-            }
-        } else {
-            emit(gen, "(");
-            emit_expr(gen, node->as.binary.left);
-            emit(gen, " %s ", binary_op_str(node->as.binary.op));
-            emit_expr(gen, node->as.binary.right);
-            emit(gen, ")");
-        }
-        break;
-
-    case NODE_UNARY:
-        emit(gen, "(%s", unary_op_str(node->as.unary.op));
-        emit_expr(gen, node->as.unary.operand);
-        emit(gen, ")");
-        break;
-
-    case NODE_CALL:
-        emit_call_expr(gen, node);
-        break;
-
-    case NODE_INDEX:
-        emit_index_expr(gen, node);
-        break;
-
-    case NODE_SLICE:
-        emit_slice_expr(gen, node);
-        break;
-
-    case NODE_MEMBER:
-        emit_member_expr(gen, node);
-        break;
-
-    case NODE_ASSIGN:
-        emit(gen, "(");
-        emit_expr(gen, node->as.assign.target);
-        emit(gen, " %s ", assign_op_str(node->as.assign.op));
-        emit_expr(gen, node->as.assign.value);
-        emit(gen, ")");
-        break;
-
-    case NODE_STRUCT_INIT:
-        emit_struct_init(gen, node);
-        break;
-
-    case NODE_TUPLE_LIT:
-        // Tuple literal: (e1, e2, ...) -> {e1, e2, ...}
-        emit(gen, "{");
-        for (int i = 0; i < node->as.tuple_lit.elements.count; i++) {
-            if (i > 0)
-                emit(gen, ", ");
-            emit_expr(gen, node->as.tuple_lit.elements.nodes[i]);
-        }
-        emit(gen, "}");
-        break;
-
-    case NODE_ARRAY_LIT:
-        // Array literal: [e1, e2, ...] -> {e1, e2, ...}
-        emit(gen, "{");
-        for (int i = 0; i < node->as.array_lit.elements.count; i++) {
-            if (i > 0)
-                emit(gen, ", ");
-            emit_expr(gen, node->as.array_lit.elements.nodes[i]);
-        }
-        emit(gen, "}");
-        break;
-
-    case NODE_NEW_EXPR:
-        emit_new_expr(gen, node);
-        break;
-
-    case NODE_STRING_INTERP:
-        emit_string_interp(gen, node);
-        break;
-
-    case NODE_CAST:
-        emit(gen, "((");
-        emit_resolved_type(gen, node->as.cast_expr.resolved_type);
-        emit(gen, ")(");
-        emit_expr(gen, node->as.cast_expr.expr);
-        emit(gen, "))");
-        break;
-
-    default:
-        emit(gen, "/* unknown expr %d */", node->type);
-        break;
-    }
-}
-
-// Emit a struct initializer as a C designated initializer: {.field = value, ...}
-static void emit_struct_init(CodeGen* gen, Node* node) {
-    emit(gen, "{");
-    for (int i = 0; i < node->as.struct_init.fields.count; i++) {
-        Node* field = node->as.struct_init.fields.nodes[i];
-        if (!field || field->type != NODE_FIELD_INIT) {
-            continue;
-        }
-        if (i > 0) {
-            emit(gen, ", ");
-        }
-        emit(gen, ".%s = ", field->as.field_init.name);
-        emit_expr(gen, field->as.field_init.value);
-    }
-    emit(gen, "}");
-}
-
-// Emit code to extract values from a tuple into variables (recursive for nested patterns)
-// temp_prefix is the expression to access the current tuple (e.g., "__tuple0" or "__tuple0._1")
-static void emit_destruct_pattern(CodeGen* gen, DestructPattern* pattern, const char* temp_prefix,
-                                  int is_const) {
-    if (!pattern)
-        return;
-
-    Type* type = pattern->resolved_type;
-
-    switch (pattern->kind) {
-    case PATTERN_IDENT:
-        // Emit: [const] Type name = temp_prefix;
-        emit_indent(gen);
-        if (is_const) {
-            emit(gen, "const ");
-        }
-        emit_resolved_type(gen, type);
-        emit(gen, " %s = %s;\n", pattern->as.ident.name, temp_prefix);
-        break;
-
-    case PATTERN_TUPLE:
-        // For tuple patterns, we need to access each element
-        // The tuple value is at temp_prefix, elements are temp_prefix._0, temp_prefix._1, etc.
-        for (int i = 0; i < pattern->as.tuple.count; i++) {
-            DestructPattern* elem = pattern->as.tuple.elements[i];
-
-            // Build the accessor string for this element
-            char accessor[256];
-            snprintf(accessor, sizeof(accessor), "%s._%d", temp_prefix, i);
-
-            if (elem->kind == PATTERN_TUPLE) {
-                // For nested tuple patterns, first create a temp variable for this level
-                Type* elem_type = elem->resolved_type;
-                emit_indent(gen);
-                emit_resolved_type(gen, elem_type);
-                int temp_id = gen->temp_count++;
-                emit(gen, " __tuple%d = %s;\n", temp_id, accessor);
-
-                // Then recursively emit the nested pattern
-                char nested_prefix[64];
-                snprintf(nested_prefix, sizeof(nested_prefix), "__tuple%d", temp_id);
-                emit_destruct_pattern(gen, elem, nested_prefix, is_const);
-            } else {
-                // Simple identifier - directly assign from accessor
-                emit_destruct_pattern(gen, elem, accessor, is_const);
-            }
-        }
-        break;
-    }
-}
-
-// Emit statements within a block without emitting braces, but with RC scope handling
-static void emit_block_contents(CodeGen* gen, Node* block) {
-    if (!block || block->type != NODE_BLOCK)
-        return;
-
-    gen->rc_scope_depth++;
-    for (int i = 0; i < block->as.block.stmts.count; i++) {
-        emit_stmt(gen, block->as.block.stmts.nodes[i]);
-    }
-    rc_cleanup_scope(gen, gen->rc_scope_depth);
-    gen->rc_scope_depth--;
-}
-
-// Emit an expression statement (NODE_EXPR_STMT).
-// Handles RC var reassignment, RC member assignment, Vec index assignment,
-// and simple expression statements.
-static void emit_expr_stmt(CodeGen* gen, Node* node) {
-    Node* expr = node->as.expr_stmt.expr;
-
-    // Handle RC reassignment: p = new_value
-    // Must inc new value, dec old value, then assign
-    if (expr->type == NODE_ASSIGN && expr->as.assign.op == TOK_EQ &&
-        expr->as.assign.target->type == NODE_IDENT &&
-        rc_is_tracked(gen, expr->as.assign.target->as.ident.name)) {
-        const char* var_name = expr->as.assign.target->as.ident.name;
-        Type*       var_type = rc_get_var_type(gen, var_name);
-        if (var_type && var_type->kind == TYPE_ENUM && var_type->as.enm.has_rc_fields) {
-            // Enum value reassignment: dec old payload, then assign, then inc if copying
-            int temp_id = gen->temp_count++;
-            emit_indent(gen);
-            emit(gen, "%s __rc_tmp%d = ", var_type->as.enm.name, temp_id);
-            emit_expr(gen, expr->as.assign.value);
-            emit(gen, ";\n");
-
-            int needs_inc = (expr->as.assign.value->type == NODE_IDENT &&
-                             rc_is_tracked(gen, expr->as.assign.value->as.ident.name));
-            if (needs_inc) {
-                emit_indent(gen);
-                emit(gen, "__rc_inc_%s(__rc_tmp%d);\n", var_type->as.enm.name, temp_id);
-            }
-
-            emit_indent(gen);
-            emit(gen, "__rc_dec_%s(%s);\n", var_type->as.enm.name, var_name);
-            emit_indent(gen);
-            emit(gen, "%s = __rc_tmp%d;\n", var_name, temp_id);
-            return;
-        }
-
-        // Evaluate new value into a temp (in case it references the old value)
-        int temp_id = gen->temp_count++;
-        emit_indent(gen);
-        emit(gen, "void* __rc_tmp%d = (void*)", temp_id);
-        emit_expr(gen, expr->as.assign.value);
-        emit(gen, ";\n");
-        emit_indent(gen);
-        emit(gen, "__rc_inc(__rc_tmp%d);\n", temp_id);
-        emit_indent(gen);
-        emit(gen, "%s(%s);\n", rc_get_dec_func(gen, var_name), var_name);
-        emit_indent(gen);
-        emit(gen, "%s = __rc_tmp%d;\n", var_name, temp_id);
-        return;
-    }
-    // Handle RC member assignment: line1.start = z
-    if (expr->type == NODE_ASSIGN && expr->as.assign.op == TOK_EQ &&
-        expr->as.assign.target->type == NODE_MEMBER) {
-        Node* member    = expr->as.assign.target;
-        int   obj_is_rc = member->as.member.object->type == NODE_IDENT &&
-                        rc_is_tracked(gen, member->as.member.object->as.ident.name);
-        if (obj_is_rc) {
-            const char* obj_name = member->as.member.object->as.ident.name;
-            Type*       obj_type = rc_get_var_type(gen, obj_name);
-            Type*       field_ty = NULL;
-            if (obj_type && obj_type->kind == TYPE_STRUCT) {
-                for (int f = 0; f < obj_type->as.struc.field_count; f++) {
-                    if (strcmp(obj_type->as.struc.field_names[f], member->as.member.name) == 0) {
-                        field_ty = obj_type->as.struc.field_types[f];
-                        break;
-                    }
-                }
-            }
-            if (field_ty && field_ty->kind == TYPE_ENUM && field_ty->as.enm.has_rc_fields) {
-                int temp_id = gen->temp_count++;
-                emit_indent(gen);
-                emit(gen, "%s __rc_tmp%d = ", field_ty->as.enm.name, temp_id);
-                emit_expr(gen, expr->as.assign.value);
-                emit(gen, ";\n");
-
-                int needs_inc = (expr->as.assign.value->type == NODE_IDENT &&
-                                 rc_is_tracked(gen, expr->as.assign.value->as.ident.name));
-                if (needs_inc) {
-                    emit_indent(gen);
-                    emit(gen, "__rc_inc_%s(__rc_tmp%d);\n", field_ty->as.enm.name, temp_id);
-                }
-
-                emit_indent(gen);
-                emit(gen, "__rc_dec_%s(", field_ty->as.enm.name);
-                emit_expr(gen, member);
-                emit(gen, ");\n");
-                emit_indent(gen);
-                emit_expr(gen, member);
-                emit(gen, " = __rc_tmp%d;\n", temp_id);
-                return;
-            }
-
-            int value_is_rc = (expr->as.assign.value->type == NODE_IDENT &&
-                               rc_is_tracked(gen, expr->as.assign.value->as.ident.name)) ||
-                              expr->as.assign.value->type == NODE_NEW_EXPR;
-            if (value_is_rc && field_ty && field_ty->kind == TYPE_STRUCT) {
-                // Determine the dec function for the field's type
-                char*       member_dec_owned = (char*)get_dec_func_for_type(field_ty);
-                const char* member_dec       = member_dec_owned;
-                int         tmp              = gen->temp_count++;
-                emit_indent(gen);
-                emit(gen, "void* __rc_tmp%d = (void*)", tmp);
-                emit_expr(gen, expr->as.assign.value);
-                emit(gen, ";\n");
-                emit_indent(gen);
-                emit(gen, "__rc_inc(__rc_tmp%d);\n", tmp);
-                emit_indent(gen);
-                emit(gen, "%s(", member_dec);
-                emit_expr(gen, member);
-                emit(gen, ");\n");
-                emit_indent(gen);
-                emit_expr(gen, member);
-                emit(gen, " = __rc_tmp%d;\n", tmp);
-                free(member_dec_owned);
-                return;
-            }
-        }
-        // Handle self.field = new_value in method bodies (self is not RC-tracked)
-        if (!obj_is_rc && member->as.member.object->type == NODE_IDENT &&
-            strcmp(member->as.member.object->as.ident.name, "self") == 0) {
-            int   value_is_rc = expr->as.assign.value->type == NODE_NEW_EXPR;
-            char* dec_fn      = NULL;
-            if (value_is_rc && gen->current_generic_template) {
-                // Generic method: look up field type from template
-                Node* ftype = lookup_generic_template_field_type(gen, member->as.member.name);
-                dec_fn      = resolve_generic_field_dec_func(gen, ftype);
-            }
-            if (dec_fn) {
-                // Dec the old field value, then assign the new one
-                emit_indent(gen);
-                emit(gen, "%s(", dec_fn);
-                emit_expr(gen, member);
-                emit(gen, ");\n");
-                emit_indent(gen);
-                emit_expr(gen, member);
-                emit(gen, " = ");
-                emit_expr(gen, expr->as.assign.value);
-                emit(gen, ";\n");
-                free(dec_fn);
-                return;
-            }
-        }
-    }
-    // Handle Vec index assignment: v[i] = x → bounds check + direct data write
-    if (expr->type == NODE_ASSIGN && expr->as.assign.target->type == NODE_INDEX &&
-        expr->as.assign.target->as.index.is_vec_index) {
-        Node* idx_node = expr->as.assign.target;
-        emit_indent(gen);
-        emit(gen, "__w0_vec_check(");
-        emit_expr(gen, idx_node->as.index.object);
-        emit(gen, "->count, ");
-        emit_expr(gen, idx_node->as.index.index);
-        emit(gen, ", %d, %d);\n", idx_node->line, idx_node->column);
-        emit_indent(gen);
-        emit_expr(gen, idx_node->as.index.object);
-        emit(gen, "->data[");
-        emit_expr(gen, idx_node->as.index.index);
-        emit(gen, "] %s ", assign_op_str(expr->as.assign.op));
-        emit_expr(gen, expr->as.assign.value);
-        emit(gen, ";\n");
-        return;
-    }
-    emit_indent(gen);
-    emit_expr(gen, expr);
-    emit(gen, ";\n");
-}
-
-// Emit an RC-managed Vec declaration: var v = new Vec<T>{...}
-static void emit_var_decl_rc_new_vec(CodeGen* gen, Node* node) {
-    Type*       rtype      = node->as.var_decl.init->as.new_expr.resolved_type;
-    const char* elem_tname = type_mangle_name(rtype->as.vec.elem);
-    emit_indent(gen);
-    emit(gen, "__Vec_%s* %s = (__Vec_%s*)__rc_alloc(sizeof(__Vec_%s));\n", elem_tname,
-         node->as.var_decl.name, elem_tname, elem_tname);
-    emit_indent(gen);
-    emit(gen, "%s->data = NULL; %s->count = 0; %s->capacity = 0;\n", node->as.var_decl.name,
-         node->as.var_decl.name, node->as.var_decl.name);
-    Node* init = node->as.var_decl.init->as.new_expr.init;
-    for (int i = 0; i < init->as.struct_init.fields.count; i++) {
-        Node* field = init->as.struct_init.fields.nodes[i];
-        if (field && field->type == NODE_FIELD_INIT) {
-            emit_indent(gen);
-            emit(gen, "__Vec_%s_push(%s, ", elem_tname, node->as.var_decl.name);
-            emit_expr(gen, field->as.field_init.value);
-            emit(gen, ");\n");
-        }
-    }
-    char dec_buf[256];
-    snprintf(dec_buf, sizeof(dec_buf), "__rc_dec_Vec_%s", elem_tname);
-    rc_push_var(gen, node->as.var_decl.name, dec_buf, rtype);
-}
-
-// Emit an RC-managed struct declaration: var p = new Point { x: 1, y: 2 }
-static void emit_var_decl_rc_new_struct(CodeGen* gen, Node* node) {
-    Type*       rtype = node->as.var_decl.init->as.new_expr.resolved_type;
-    const char* tname = rtype->as.struc.name;
-    emit_indent(gen);
-    emit(gen, "%s* %s = (%s*)__rc_alloc(sizeof(%s));\n", tname, node->as.var_decl.name, tname,
-         tname);
-    emit_indent(gen);
-    emit(gen, "*%s = (%s)", node->as.var_decl.name, tname);
-    emit_struct_init(gen, node->as.var_decl.init->as.new_expr.init);
-    emit(gen, ";\n");
-    // Increment refcount for any RC-tracked idents stored in struct fields
-    Node* rc_init = node->as.var_decl.init->as.new_expr.init;
-    for (int i = 0; i < rc_init->as.struct_init.fields.count; i++) {
-        Node* field = rc_init->as.struct_init.fields.nodes[i];
-        if (field && field->type == NODE_FIELD_INIT &&
-            field->as.field_init.value->type == NODE_IDENT &&
-            rc_is_tracked(gen, field->as.field_init.value->as.ident.name)) {
-            const char* vname  = field->as.field_init.value->as.ident.name;
-            Type*       vtype  = rc_get_var_type(gen, vname);
-            const char* inc_fn = get_inc_func_for_type(vtype);
-            emit_indent(gen);
-            emit(gen, "%s(%s);\n", inc_fn, vname);
-            free((char*)inc_fn);
-        }
-    }
-    const char* dec_fn = get_dec_func_for_type(rtype);
-    rc_push_var(gen, node->as.var_decl.name, dec_fn, rtype);
-    free((char*)dec_fn);
-}
-
-// Emit an RC copy or ownership transfer: var x = existing_rc_var or func_call()
-static void emit_var_decl_rc_copy(CodeGen* gen, Node* node) {
-    emit_indent(gen);
-    if (node->as.var_decl.type) {
-        emit_type_with_name(gen, node->as.var_decl.type, node->as.var_decl.name);
-    } else if (node->as.var_decl.resolved_type) {
-        Type* rtype = node->as.var_decl.resolved_type;
-        if (rtype->kind == TYPE_STRUCT) {
-            emit(gen, "%s* %s", rtype->as.struc.name, node->as.var_decl.name);
-        } else {
-            emit_resolved_type(gen, rtype);
-            emit(gen, " %s", node->as.var_decl.name);
-        }
-    } else {
-        emit(gen, "void* %s", node->as.var_decl.name);
-    }
-    emit(gen, " = ");
-    emit_expr(gen, node->as.var_decl.init);
-    emit(gen, ";\n");
-    // Function calls transfer ownership (rc already 1), no inc needed
-    Type* rc_type = node->as.var_decl.resolved_type;
-    int   skip_inc =
-        node->as.var_decl.init->type == NODE_CALL ||
-        (rc_type && rc_type->kind == TYPE_ENUM && node->as.var_decl.init->type == NODE_ENUM_VALUE);
-    if (!skip_inc && rc_type) {
-        const char* inc_fn = get_inc_func_for_type(rc_type);
-        emit_indent(gen);
-        emit(gen, "%s(%s);\n", inc_fn, node->as.var_decl.name);
-        free((char*)inc_fn);
-    }
-    const char* dec_fn = get_dec_func_for_type(rc_type);
-    rc_push_var(gen, node->as.var_decl.name, dec_fn, rc_type);
-    free((char*)dec_fn);
-}
-
-// Emit a type-and-name declaration inferred from the initializer expression.
-static void emit_var_decl_inferred_type(CodeGen* gen, Node* node) {
-    if (!node->as.var_decl.init) {
-        emit(gen, "int64_t %s", node->as.var_decl.name);
-        return;
-    }
-    switch (node->as.var_decl.init->type) {
-    case NODE_INT_LIT:
-        emit(gen, "int64_t %s", node->as.var_decl.name);
-        break;
-    case NODE_FLOAT_LIT:
-        emit(gen, "float %s", node->as.var_decl.name);
-        break;
-    case NODE_BOOL_LIT:
-        emit(gen, "bool %s", node->as.var_decl.name);
-        break;
-    case NODE_STRING_LIT:
-        emit(gen, "const char* %s", node->as.var_decl.name);
-        break;
-    case NODE_CHAR_LIT:
-        emit(gen, "char %s", node->as.var_decl.name);
-        break;
-    case NODE_TUPLE_LIT: {
-        int    count = node->as.var_decl.init->as.tuple_lit.elements.count;
-        Type** elems = xmalloc(count * sizeof(Type*));
-        for (int i = 0; i < count; i++) {
-            Node* elem = node->as.var_decl.init->as.tuple_lit.elements.nodes[i];
-            switch (elem->type) {
-            case NODE_INT_LIT:
-                elems[i] = type_int64;
-                break;
-            case NODE_FLOAT_LIT:
-                elems[i] = type_f32;
-                break;
-            case NODE_BOOL_LIT:
-                elems[i] = type_bool;
-                break;
-            case NODE_STRING_LIT:
-                elems[i] = type_string;
-                break;
-            case NODE_CHAR_LIT:
-                elems[i] = type_char;
-                break;
-            case NODE_TUPLE_LIT:
-                elems[i] = type_int64; // Fallback
-                break;
-            default:
-                elems[i] = type_int64; // Default
-                break;
-            }
-        }
-        Type* tuple = type_tuple(elems, count);
-        int   idx   = -1;
-        for (int i = 0; i < gen->tuple_type_count; i++) {
-            if (tuple_types_equal(gen->tuple_types[i], tuple)) {
-                idx = i;
-                break;
-            }
-        }
-        if (idx >= 0) {
-            emit(gen, "__tuple_t%d %s", idx, node->as.var_decl.name);
-        } else {
-            emit(gen, "struct { ");
-            for (int i = 0; i < count; i++) {
-                emit_resolved_type(gen, elems[i]);
-                emit(gen, " _%d; ", i);
-            }
-            emit(gen, "} %s", node->as.var_decl.name);
-        }
-        break;
-    }
-    case NODE_ARRAY_LIT: {
-        Node* init      = node->as.var_decl.init;
-        Type* elem_type = init->as.array_lit.resolved_type;
-        int   count     = init->as.array_lit.elements.count;
-        emit_resolved_type(gen, elem_type);
-        emit(gen, " %s[%d]", node->as.var_decl.name, count);
-        break;
-    }
-    case NODE_ENUM_VALUE:
-        emit(gen, "%.*s %s", node->as.var_decl.init->as.enum_value.enum_name_length,
-             node->as.var_decl.init->as.enum_value.enum_name, node->as.var_decl.name);
-        break;
-    default:
-        if (node->as.var_decl.resolved_type) {
-            emit_resolved_type(gen, node->as.var_decl.resolved_type);
-            emit(gen, " %s", node->as.var_decl.name);
-        } else {
-            emit(gen, "int64_t %s", node->as.var_decl.name);
-        }
-        break;
-    }
-}
-
-// Emit a variable declaration statement (NODE_VAR_DECL).
-// Handles destructuring, RC-managed declarations, type inference, and struct init.
-static void emit_var_decl_stmt(CodeGen* gen, Node* node) {
-    // Handle destructuring: var (a, b) = tuple; or var (a, (b, c)) = nested;
-    DestructPattern* pattern = node->as.var_decl.destruct_pattern;
-    if (pattern) {
-        Type* tuple_type = pattern->resolved_type;
-        emit_indent(gen);
-        emit_resolved_type(gen, tuple_type);
-        int temp_id = gen->temp_count++;
-        emit(gen, " __tuple%d = ", temp_id);
-        emit_expr(gen, node->as.var_decl.init);
-        emit(gen, ";\n");
-        char temp_prefix[64];
-        snprintf(temp_prefix, sizeof(temp_prefix), "__tuple%d", temp_id);
-        emit_destruct_pattern(gen, pattern, temp_prefix, node->as.var_decl.is_const);
-        return;
-    }
-
-    // Handle RC-managed variable declarations
-    if (node->as.var_decl.is_rc && node->as.var_decl.init) {
-        if (node->as.var_decl.init->type == NODE_NEW_EXPR) {
-            Type* rtype = node->as.var_decl.init->as.new_expr.resolved_type;
-            if (rtype && rtype->kind == TYPE_VEC) {
-                emit_var_decl_rc_new_vec(gen, node);
-            } else {
-                emit_var_decl_rc_new_struct(gen, node);
-            }
-        } else {
-            emit_var_decl_rc_copy(gen, node);
-        }
-        return;
-    }
-
-    emit_indent(gen);
-    if (node->as.var_decl.is_const) {
-        emit(gen, "const ");
-    }
-
-    int struct_type = node->as.var_decl.type && is_struct_type(gen, node->as.var_decl.type);
-
-    if (node->as.var_decl.type) {
-        emit_type_with_name(gen, node->as.var_decl.type, node->as.var_decl.name);
-    } else {
-        emit_var_decl_inferred_type(gen, node);
-    }
-
-    if (node->as.var_decl.init) {
-        if (struct_type && node->as.var_decl.init->type == NODE_NULL_LIT) {
-            emit(gen, " = NULL");
-        } else {
-            emit(gen, " = ");
-            emit_expr(gen, node->as.var_decl.init);
-        }
-    }
-    emit(gen, ";\n");
-}
-
-// Emit a return statement (NODE_RETURN).
-// Handles RC cleanup, defer integration, and return value evaluation.
-static void emit_return_stmt(CodeGen* gen, Node* node) {
-    // Determine if we're returning an RC var (skip it in cleanup)
-    const char* skip_name = NULL;
-    if (node->as.return_stmt.value && node->as.return_stmt.value->type == NODE_IDENT) {
-        // Check if the returned identifier is an RC var
-        const char* ret_name = node->as.return_stmt.value->as.ident.name;
-        for (int i = 0; i < gen->rc_var_count; i++) {
-            if (strcmp(gen->rc_vars[i].name, ret_name) == 0) {
-                skip_name = ret_name;
-                break;
-            }
-        }
-    }
-
-    if (gen->defer_count > 0) {
-        // With defers: store value in __ret, cleanup RC, goto cleanup
-        emit_indent(gen);
-        if (node->as.return_stmt.value) {
-            emit(gen, "__ret = ");
-            emit_expr(gen, node->as.return_stmt.value);
-            emit(gen, ";\n");
-        }
-        if (gen->rc_var_count > 0) {
-            rc_cleanup_all(gen, skip_name);
-        }
-        emit_indent(gen);
-        emit(gen, "goto __cleanup;\n");
-    } else {
-        // No defers: cleanup RC, then return
-        if (gen->rc_var_count > 0) {
-            if (node->as.return_stmt.value && !skip_name) {
-                // Complex expression: evaluate to temp first
-                emit_indent(gen);
-                emit(gen, "typeof(");
-                emit_expr(gen, node->as.return_stmt.value);
-                emit(gen, ") __rc_ret = ");
-                emit_expr(gen, node->as.return_stmt.value);
-                emit(gen, ";\n");
-                rc_cleanup_all(gen, NULL);
-                emit_indent(gen);
-                emit(gen, "return __rc_ret;\n");
-            } else {
-                rc_cleanup_all(gen, skip_name);
-                emit_indent(gen);
-                emit(gen, "return");
-                if (node->as.return_stmt.value) {
-                    emit(gen, " ");
-                    emit_expr(gen, node->as.return_stmt.value);
-                }
-                emit(gen, ";\n");
-            }
-        } else {
-            emit_indent(gen);
-            emit(gen, "return");
-            if (node->as.return_stmt.value) {
-                emit(gen, " ");
-                emit_expr(gen, node->as.return_stmt.value);
-            }
-            emit(gen, ";\n");
-        }
-    }
-}
-
-// Emit a match statement as an if/else-if chain over the enum tag
-static void emit_match_stmt(CodeGen* gen, Node* node) {
-    Type* enum_type = node->as.match_stmt.resolved_type;
-    if (!enum_type)
-        return;
-
-    int         is_data   = enum_type->as.enm.has_data;
-    const char* enum_name = enum_type->as.enm.name;
-
-    // Emit temp variable: EnumType __matchN = <expr>;
-    int match_id = gen->temp_count++;
-    emit_indent(gen);
-    emit(gen, "%s __match%d = ", enum_name, match_id);
-    emit_expr(gen, node->as.match_stmt.expr);
-    emit(gen, ";\n");
-
-    int first = 1;
-    for (int a = 0; a < node->as.match_stmt.arms.count; a++) {
-        Node* arm = node->as.match_stmt.arms.nodes[a];
-
-        emit_indent(gen);
-        if (arm->as.match_arm.is_wildcard) {
-            if (first) {
-                emit(gen, "{\n");
-            } else {
-                emit(gen, "else {\n");
-            }
-        } else {
-            const char* variant = arm->as.match_arm.variant_name;
-            if (first) {
-                emit(gen, "if (");
-            } else {
-                emit(gen, "else if (");
-            }
-            if (is_data) {
-                emit(gen, "__match%d.tag == %s_%s", match_id, enum_name, variant);
-            } else {
-                emit(gen, "__match%d == %s_%s", match_id, enum_name, variant);
-            }
-            emit(gen, ") {\n");
-        }
-        first = 0;
-
-        gen->indent++;
-
-        // Emit binding declarations for data enum variants
-        if (!arm->as.match_arm.is_wildcard && is_data && arm->as.match_arm.binding_count > 0) {
-            const char* variant = arm->as.match_arm.variant_name;
-            // Look up variant index to get types
-            int variant_idx = -1;
-            for (int i = 0; i < enum_type->as.enm.value_count; i++) {
-                if (strcmp(enum_type->as.enm.value_names[i], variant) == 0) {
-                    variant_idx = i;
-                    break;
-                }
-            }
-            for (int j = 0; j < arm->as.match_arm.binding_count; j++) {
-                emit_indent(gen);
-                emit_resolved_type(gen, enum_type->as.enm.variant_types[variant_idx][j]);
-                emit(gen, " %s = __match%d.%s.f%d;\n", arm->as.match_arm.bindings[j], match_id,
-                     variant, j);
-            }
-        }
-
-        // Emit arm body
-        if (arm->as.match_arm.body) {
-            if (arm->as.match_arm.body->type == NODE_BLOCK) {
-                emit_block_contents(gen, arm->as.match_arm.body);
-            } else {
-                emit_stmt(gen, arm->as.match_arm.body);
-            }
-        }
-
-        gen->indent--;
-        emit_indent(gen);
-        emit(gen, "}\n");
-    }
-}
-
-// Dispatch statement code generation based on node type
-void emit_stmt(CodeGen* gen, Node* node) {
-    if (!node)
-        return;
-
-    switch (node->type) {
-    case NODE_EXPR_STMT:
-        emit_expr_stmt(gen, node);
-        break;
-
-    case NODE_VAR_DECL:
-        emit_var_decl_stmt(gen, node);
-        break;
-
-    case NODE_BLOCK:
-        emit_indent(gen);
-        emit(gen, "{\n");
-        gen->indent++;
-        gen->rc_scope_depth++;
-        for (int i = 0; i < node->as.block.stmts.count; i++) {
-            emit_stmt(gen, node->as.block.stmts.nodes[i]);
-        }
-        rc_cleanup_scope(gen, gen->rc_scope_depth);
-        gen->rc_scope_depth--;
-        gen->indent--;
-        emit_indent(gen);
-        emit(gen, "}\n");
-        break;
-
-    case NODE_IF: {
-        emit_indent(gen);
-        int cond_has_parens = (node->as.if_stmt.cond->type == NODE_BINARY ||
-                               node->as.if_stmt.cond->type == NODE_UNARY);
-        if (cond_has_parens) {
-            emit(gen, "if ");
-            emit_expr(gen, node->as.if_stmt.cond);
-            emit(gen, " {\n");
-        } else {
-            emit(gen, "if (");
-            emit_expr(gen, node->as.if_stmt.cond);
-            emit(gen, ") {\n");
-        }
-        gen->indent++;
-        // Emit then block contents directly (it's already a block)
-        if (node->as.if_stmt.then_block->type == NODE_BLOCK) {
-            emit_block_contents(gen, node->as.if_stmt.then_block);
-        } else {
-            emit_stmt(gen, node->as.if_stmt.then_block);
-        }
-        gen->indent--;
-        emit_indent(gen);
-        emit(gen, "}");
-
-        if (node->as.if_stmt.else_block) {
-            if (node->as.if_stmt.else_block->type == NODE_IF) {
-                // else if
-                emit(gen, " else ");
-                // Remove indent for else if
-                gen->indent--;
-                emit_stmt(gen, node->as.if_stmt.else_block);
-                gen->indent++;
-            } else {
-                emit(gen, " else {\n");
-                gen->indent++;
-                if (node->as.if_stmt.else_block->type == NODE_BLOCK) {
-                    emit_block_contents(gen, node->as.if_stmt.else_block);
-                } else {
-                    emit_stmt(gen, node->as.if_stmt.else_block);
-                }
-                gen->indent--;
-                emit_indent(gen);
-                emit(gen, "}\n");
-            }
-        } else {
-            emit(gen, "\n");
-        }
-        break;
-    }
-
-    case NODE_WHILE: {
-        emit_indent(gen);
-        int wcond_has_parens = (node->as.while_stmt.cond->type == NODE_BINARY ||
-                                node->as.while_stmt.cond->type == NODE_UNARY);
-        if (wcond_has_parens) {
-            emit(gen, "while ");
-            emit_expr(gen, node->as.while_stmt.cond);
-            emit(gen, " {\n");
-        } else {
-            emit(gen, "while (");
-            emit_expr(gen, node->as.while_stmt.cond);
-            emit(gen, ") {\n");
-        }
-        gen->indent++;
-        if (node->as.while_stmt.body->type == NODE_BLOCK) {
-            emit_block_contents(gen, node->as.while_stmt.body);
-        } else {
-            emit_stmt(gen, node->as.while_stmt.body);
-        }
-        gen->indent--;
-        emit_indent(gen);
-        emit(gen, "}\n");
-        break;
-    }
-
-    case NODE_FOR:
-        emit_indent(gen);
-        emit(gen, "for (");
-        // Init
-        if (node->as.for_stmt.init) {
-            if (node->as.for_stmt.init->type == NODE_VAR_DECL) {
-                Node* v = node->as.for_stmt.init;
-                if (v->as.var_decl.type) {
-                    emit_type_with_name(gen, v->as.var_decl.type, v->as.var_decl.name);
-                } else {
-                    emit(gen, "int64_t %s", v->as.var_decl.name);
-                }
-                if (v->as.var_decl.init) {
-                    emit(gen, " = ");
-                    emit_expr(gen, v->as.var_decl.init);
-                }
-            } else {
-                emit_expr(gen, node->as.for_stmt.init);
-            }
-        }
-        emit(gen, "; ");
-        // Cond
-        if (node->as.for_stmt.cond) {
-            emit_expr(gen, node->as.for_stmt.cond);
-        }
-        emit(gen, "; ");
-        // Post
-        if (node->as.for_stmt.post) {
-            emit_expr(gen, node->as.for_stmt.post);
-        }
-        emit(gen, ") {\n");
-        gen->indent++;
-        if (node->as.for_stmt.body->type == NODE_BLOCK) {
-            emit_block_contents(gen, node->as.for_stmt.body);
-        } else {
-            emit_stmt(gen, node->as.for_stmt.body);
-        }
-        gen->indent--;
-        emit_indent(gen);
-        emit(gen, "}\n");
-        break;
-
-    case NODE_FOREACH:
-        if (node->as.foreach_stmt.collection) {
-            int idx_id = gen->temp_count++;
-            if (node->as.foreach_stmt.is_string) {
-                // String foreach: foreach (const c in str)
-                emit_indent(gen);
-                emit(gen, "for (int64_t __foreach_%d = 0; __foreach_%d < (int64_t)strlen(", idx_id,
-                     idx_id);
-                emit_expr(gen, node->as.foreach_stmt.collection);
-                emit(gen, "); __foreach_%d++) {\n", idx_id);
-                gen->indent++;
-                emit_indent(gen);
-                emit(gen, "char %s = ", node->as.foreach_stmt.var_name);
-                emit_expr(gen, node->as.foreach_stmt.collection);
-                emit(gen, "[__foreach_%d];\n", idx_id);
-            } else {
-                // Collection foreach: foreach (const item in collection)
-                // Vec uses -> (pointer), Span uses . (value type)
-                const char* access = node->as.foreach_stmt.is_span ? "." : "->";
-                emit_indent(gen);
-                emit(gen, "for (int64_t __foreach_%d = 0; __foreach_%d < ", idx_id, idx_id);
-                emit_expr(gen, node->as.foreach_stmt.collection);
-                emit(gen, "%scount; __foreach_%d++) {\n", access, idx_id);
-                gen->indent++;
-                // Declare the loop variable from collection data[idx]
-                emit_indent(gen);
-                emit_resolved_type(gen, node->as.foreach_stmt.resolved_type
-                                            ? node->as.foreach_stmt.resolved_type
-                                            : type_int64);
-                emit(gen, " %s = ", node->as.foreach_stmt.var_name);
-                emit_expr(gen, node->as.foreach_stmt.collection);
-                emit(gen, "%sdata[__foreach_%d];\n", access, idx_id);
-            }
-            if (node->as.foreach_stmt.body->type == NODE_BLOCK) {
-                emit_block_contents(gen, node->as.foreach_stmt.body);
-            } else {
-                emit_stmt(gen, node->as.foreach_stmt.body);
-            }
-            gen->indent--;
-            emit_indent(gen);
-            emit(gen, "}\n");
-        } else {
-            // Range foreach: for (<type> var = start; var < end; var += step)
-            emit_indent(gen);
-            emit(gen, "for (");
-            emit_resolved_type(gen, node->as.foreach_stmt.resolved_type
-                                        ? node->as.foreach_stmt.resolved_type
-                                        : type_int64);
-            emit(gen, " %s = ", node->as.foreach_stmt.var_name);
-            emit_expr(gen, node->as.foreach_stmt.start);
-            emit(gen, "; %s < ", node->as.foreach_stmt.var_name);
-            emit_expr(gen, node->as.foreach_stmt.end);
-            emit(gen, "; %s += ", node->as.foreach_stmt.var_name);
-            emit_expr(gen, node->as.foreach_stmt.step);
-            emit(gen, ") {\n");
-            gen->indent++;
-            if (node->as.foreach_stmt.body->type == NODE_BLOCK) {
-                emit_block_contents(gen, node->as.foreach_stmt.body);
-            } else {
-                emit_stmt(gen, node->as.foreach_stmt.body);
-            }
-            gen->indent--;
-            emit_indent(gen);
-            emit(gen, "}\n");
-        }
-        break;
-    case NODE_RETURN:
-        emit_return_stmt(gen, node);
-        break;
-
-    case NODE_DEFER:
-        // Don't emit anything here - just push to defer stack
-        defer_push(gen, node->as.defer_stmt.stmt);
-        break;
-
-    case NODE_MATCH:
-        emit_match_stmt(gen, node);
-        break;
-
-    case NODE_BREAK:
-        rc_cleanup_scope(gen, gen->rc_scope_depth);
-        emit_indent(gen);
-        emit(gen, "break;\n");
-        break;
-
-    case NODE_CONTINUE:
-        rc_cleanup_scope(gen, gen->rc_scope_depth);
-        emit_indent(gen);
-        emit(gen, "continue;\n");
-        break;
-
-    default:
-        emit_indent(gen);
-        emit(gen, "/* unknown stmt %d */;\n", node->type);
-        break;
     }
 }
